fix: RAPL sensor now properly use CPU share instead of full measure

+++ runtime/src/main/java/io/quarkiverse/power/runtime/sensors/linux/rapl/IntelRAPLMeasure.java
@@ -3,17 +3,17 @@ package io.quarkiverse.power.runtime.sensors.linux.rapl;
 import io.quarkiverse.power.runtime.sensors.IncrementableMeasure;
 
 public class IntelRAPLMeasure implements IncrementableMeasure {
-    private final long initial;
-    private long cpu;
+    private final double initial;
+    private double cpu;
     private final long startedAt = System.currentTimeMillis();
 
-    public IntelRAPLMeasure(long initial) {
+    public IntelRAPLMeasure(double initial) {
         this.initial = initial;
     }
 
     @Override
     public double cpu() {
-        return ((double) cpu / durationSinceStart()) / 1_000;
+        return (cpu / durationSinceStart()) / 1_000;
     }
 
     @Override
